commit 4 name error

+++ src/main/java/ru/practicum/shareit/booking/BookingRepository.java
@@ -3,59 +3,68 @@ package ru.practicum.shareit.booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import ru.practicum.shareit.booking.model.Booking;
-import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.user.User;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-public interface BookingRepository  extends JpaRepository<Booking, Integer> {
+public interface BookingRepository  extends JpaRepository<Booking, Long> {
     @Query("select b from Booking b where ( " +
             ":start <= b.start and b.start <= :end or " +
             ":start <= b.end and b.end <= :end) and " +
             "b.item.id = :itemId and " +
             "b.status = 'APPROVED'")
-    List<Booking> findAllByDateAndId(int itemId, LocalDateTime start, LocalDateTime end);
+    List<Booking> findAllByDateAndId(Long itemId, LocalDateTime start, LocalDateTime end);
 
-    List<Booking> findAllByBookerId(int bookerId);
+    //Пользователь смотрит свои заказы
+    //////////////////////////////////////////
+    List<Booking> findAllByBookerId(Long bookerId);
 
-    List<Booking> findAllByBookerIdAndStatus(int bookerId, BookingStatus status);
+    List<Booking> findAllByBookerIdAndStatus(Long bookerId, BookingStatus status);
 
-    List<Booking> findAllByBookerIdAndEndIsBefore(int bookerId, LocalDateTime date);
+    List<Booking> findAllByBookerIdAndEndIsBefore(Long bookerId, LocalDateTime date);
 
-    List<Booking> findAllByBookerIdAndStartIsAfter(int bookerId, LocalDateTime date);
+    List<Booking> findAllByBookerIdAndStartIsAfter(Long bookerId, LocalDateTime date);
 
     @Query("select b from Booking b where  " +
             "b.start < :date and :date < b.end and " +
             "b.booker.id = :bookerId")
-    List<Booking> findByBookerIdCurrDate(int bookerId, LocalDateTime date);
+    List<Booking> findByBookerIdCurrDate(Long bookerId, LocalDateTime date);
+    //////////////////////////////////////////
 
+    //Пользователь смотрит свои вещи, которые были заказаны другими.
+    //////////////////////////////////////////
     @Query("select b from Booking b where  " +
             "b.item.owner.id = :ownerId")
-    List<Booking> findAllItemBooking(int ownerId);
+    List<Booking> findAllItemBooking(Long ownerId);
 
     @Query("select b from Booking b where  " +
             "b.item.owner.id = :ownerId and " +
             "b.end < :date")
-    List<Booking> findAllItemBookingEndIsBefore(int ownerId, LocalDateTime date);
+    List<Booking> findAllItemBookingEndIsBefore(Long ownerId, LocalDateTime date);
 
     @Query("select b from Booking b where  " +
             "b.item.owner.id = :ownerId and " +
             "b.start > :date")
-    List<Booking> findAllItemBookingAndStartIsAfter(int ownerId, LocalDateTime date);
+    List<Booking> findAllItemBookingAndStartIsAfter(Long ownerId, LocalDateTime date);
 
     @Query("select b from Booking b where  " +
             "b.start < :date and :date < b.end and " +
             "b.item.owner.id = :ownerId")
-    List<Booking> findAllItemBookingCurrDate(int ownerId, LocalDateTime date);
+    List<Booking> findAllItemBookingCurrDate(Long ownerId, LocalDateTime date);
 
     @Query("select b from Booking b where  " +
             "b.item.owner.id = :ownerId and b.status = :status")
-    List<Booking> findAllItemBookingStatus(int ownerId, BookingStatus status);
+    List<Booking> findAllItemBookingStatus(Long ownerId, BookingStatus status);
+    //////////////////////////////////////////
 
-    Optional<Booking> findFirstByItemIdAndEndBeforeOrderByEndDesc(int itemId, LocalDateTime date);
+    //дата последнего бронирования
+    Optional<Booking> findFirstByItemIdAndEndBeforeOrderByEndDesc(Long itemId, LocalDateTime date);
 
-    Optional<Booking> findFirstByItemIdAndStartAfterOrderByStart(int itemId, LocalDateTime date);
+    //дата следующего бронирования
+    Optional<Booking> findFirstByItemIdAndStartAfterOrderByStart(Long itemId, LocalDateTime date);
 
-    Optional<Booking> findFirstByBookerAndItemIdAndEndBefore(User booker, int itemId, LocalDateTime date);
+    //Пользователь хотя бы раз бронировал вещь
+    Optional<Booking> findFirstByBookerAndItemIdAndEndBefore(User booker, Long itemId, LocalDateTime date);
 }
