+++ src/main/java/ru/practicum/shareit/item/comment/model/Comment.java
@@ -3,7 +3,7 @@ package ru.practicum.shareit.item.comment.model;
 import lombok.Getter;
 import lombok.Setter;
 import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.user.User;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
@@ -17,7 +17,7 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id")
-    private  int id;
+    private  Long id;
 
     @Column(nullable = false)
     private  String text;
